Expression template for Bayes Factor results

Expression template for Bayes Factor results

bf_expr_template(
  top.text,
  estimate.df,
  prior.type = NULL,
  estimate.type = NULL,
  centrality = "median",
  conf.level = 0.95,
  conf.method = "HDI",
  k = 2L,
  ...
)

Arguments

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

estimate.df
Dataframe containing estimates and their credible intervals along with Bayes Factor value. The columns should be named as estimate, estimate.LB, estimate.UB, and bf10.

prior.type
A character that specifies the prior type.

estimate.type
A character that specifies the relevant effect size.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

conf.method
The type of index used for Credible Interval. Can be "hdi" (default), "eti", or "si" (see si(), hdi(), eti() functions from bayestestR package).

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Currently ignored.