Making expression for contingency table and goodness of fit tests

Making expression for contingency analysis (Pearson's chi-square test for independence for between-subjects design or McNemar's test for within-subjects design) or goodness of fit test for a single categorical variable.

expr_contingency_tab(data, x, y = NULL, counts = NULL, ratio = NULL,
  nboot = 100, paired = FALSE, stat.title = NULL,
  legend.title = NULL, conf.level = 0.95, conf.type = "norm",
  simulate.p.value = FALSE, B = 2000, bias.correct = FALSE, k = 2,
  messages = TRUE, ...)

expr_onesample_proptest(data, x, y = NULL, counts = NULL,
  ratio = NULL, nboot = 100, paired = FALSE, stat.title = NULL,
  legend.title = NULL, conf.level = 0.95, conf.type = "norm",
  simulate.p.value = FALSE, B = 2000, bias.correct = FALSE, k = 2,
  messages = TRUE, ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The variable to use as the rows in the contingency table.

y
The variable to use as the columns in the contingency table. Default is NULL. If NULL, one-sample proportion test (a goodness of fit test) will be run for the main variable. Otherwise an appropriate association test will be run.

counts
A string naming a variable in data containing counts, or NULL if each row represents a single observation (Default).

ratio
A vector of proportions: the expected proportions for the proportion test (should sum to 1). Default is NULL, which means the null is equal theoretical proportions across the levels of the nominal variable. This means if there are two levels this will be ratio = c(0.5,0.5) or if there are four levels this will be ratio = c(0.25,0.25,0.25,0.25), etc.

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100).

paired
Logical indicating whether data came from a within-subjects or repeated measures design study (Default: FALSE). If TRUE, McNemar's test subtitle will be returned. If FALSE, Pearson's chi-square test will be returned.

stat.title
Title for the effect being investigated with the chi-square test. The default is NULL, i.e. no title will be added to describe the effect being shown. An example of a stat.title argument will be something like "main x condition" or "interaction".

legend.title
Title text for the legend.

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% lower and upper confidence intervals (0.95).

conf.type
A vector of character strings representing the type of intervals required. The value should be any subset of the values "norm", "basic", "perc", "bca". For more, see ?boot::boot.ci.

simulate.p.value
a logical indicating whether to compute p-values by Monte Carlo simulation.

B
an integer specifying the number of replicates used in the Monte Carlo test.

bias.correct
If TRUE, a bias correction will be applied to Cramer's V.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

...
Additional arguments (currently ignored).

Details

For more details about how the effect sizes and their confidence intervals were computed, see documentation in ?rcompanion::cramerV, ?rcompanion::cramerVFit, and ?rcompanion::cohenG.

Examples

# ------------------------ association tests -----------------------------

set.seed(123)
library(statsExpressions)

# without counts data
statsExpressions::expr_contingency_tab(
  data = mtcars,
  x = am,
  y = cyl,
  nboot = 15
)

#> Warning: Chi-squared approximation may be incorrect

#> Note: 95% CI for effect size estimate was computed with 15 bootstrap samples.
#> 

#> 

#> paste(NULL, chi["Pearson"]^2, "(", "2", ") = ", "8.74", ", ", 
#>     italic("p"), " = ", "0.013", ", ", italic("V")["Cramer"], 
#>     " = ", "0.52", ", CI"["95%"], " [", "0.20", ", ", "0.71", 
#>     "]", ", ", italic("n")["obs"], " = ", 32L)

# ------------------------ goodness of fit tests ---------------------------

# for reproducibility
set.seed(123)

# with counts
expr_contingency_tab(
  data = as.data.frame(HairEyeColor),
  x = Eye,
  counts = Freq,
  ratio = c(0.2, 0.2, 0.3, 0.3)
)

#> Note: 95% CI for effect size estimate was computed with 100 bootstrap samples.
#> 

#> 

#> paste(NULL, chi["gof"]^2, "(", "3", ") = ", "278.96", ", ", italic("p"), 
#>     " = ", "< 0.001", ", ", italic("V")["Cramer"], " = ", "0.40", 
#>     ", CI"["95%"], " [", "0.35", ", ", "0.43", "]", ", ", italic("n")["obs"], 
#>     " = ", 592L)