Expression and dataframe for correlation analysis

Expression and dataframe for correlation analysis

expr_corr_test(
  data,
  x,
  y,
  type = "parametric",
  k = 2L,
  conf.level = 0.95,
  beta = 0.1,
  bf.prior = 0.707,
  top.text = NULL,
  output = "expression",
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The column in data containing the explanatory variable to be plotted on the x-axis. Can be entered either as a character string (e.g., "x") or as a bare expression (e.g, x).

y
The column in data containing the response (outcome) variable to be plotted on the y-axis. Can be entered either as a character string (e.g., "y") or as a bare expression (e.g, y).

type
Type of association between paired samples required (""parametric": Pearson's product moment correlation coefficient" or ""nonparametric": Spearman's rho" or ""robust": percentage bend correlation coefficient" or ""bayes": Bayes Factor for Pearson's r"). Corresponding abbreviations are also accepted: "p" (for parametric/pearson), "np" (nonparametric/spearman), "r" (robust), "bf" (for bayes factor), resp.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

beta
bending constant (Default: 0.1). For more, see WRS2::pbcor().

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

...
Additional arguments (currently ignored).

Value

Expression containing results from correlation test with confidence intervals for the correlation coefficient estimate. Results are extracted via correlation::correlation.

References

For more details, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# for reproducibility
set.seed(123)
library(statsExpressions)

# without changing defaults
expr_corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack
)

#> paste(italic("t")["Student"], "(", "435", ") = ", "-0.97", ", ", 
#>     italic("p"), " = ", "0.332", ", ", widehat(italic("r"))["Pearson"], 
#>     " = ", "-0.05", ", CI"["95%"], " [", "-0.14", ", ", "0.05", 
#>     "]", ", ", italic("n")["pairs"], " = ", "437")

# changing defaults
expr_corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack,
  type = "robust",
  output = "dataframe"
)

#> # A tibble: 1 x 12
#>   parameter1 parameter2 estimate conf.low conf.high statistic    df p.value
#>   <chr>      <chr>         <dbl>    <dbl>     <dbl>     <dbl> <dbl>   <dbl>
#> 1 area       percblack   0.00240  -0.0914    0.0962    0.0500   435   0.960
#>   method          n.obs effectsize      ci.width
#>   <chr>           <int> <chr>              <dbl>
#> 1 Percentage Bend   437 Percentage Bend     0.95