Making expression for random-effects meta-analysis

Making expression for random-effects meta-analysis

expr_meta_random(
  data,
  type = "parametric",
  metaBMA.args = list(),
  random = "mixture",
  k = 2L,
  conf.level = 0.95,
  caption = NULL,
  output = "expression",
  ...
)

Arguments

data
A dataframe. It must contain columns named estimate (effect sizes or outcomes) and std.error (corresponding standard errors). These two columns will be used for yi and sei arguments in metafor::rma (for parametric analysis) or metaplus::metaplus (for robust analysis).

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

metaBMA.args
A list of additional arguments to be passed to metaBMA::meta_random.

random
The type of random effects distribution. One of "normal","t-dist","mixture", for standard normal, \(t\)-distribution or mixture of normals respectively.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

caption
Text to display as caption. This argument is relevant only when output = "caption".

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

...
Additional arguments passed to the respective meta-analysis function.

Details

This analysis is carried out using-

parametric: metafor::rma

robust: metaplus::metaplus

Bayesian: metaBMA::meta_random

Examples

# \donttest{
# setup
set.seed(123)
library(statsExpressions)
library(metaplus)

#> 
#> Note: The second parameter is the study standard error not the standard error squared as in the metafor package.

# renaming to what `statsExpressions` expects
df <- dplyr::rename(mag, estimate = yi, std.error = sei)

# ----------------------- parametric ---------------------------------------

# creating expression
expr_meta_random(data = df, k = 3)

#> paste(italic("z"), " = ", "-3.619", ", ", italic("p"), " = ", 
#>     "2.95e-04", ", ", widehat(beta)["summary"]^"meta", " = ", 
#>     "-0.767", ", CI"["95%"], " [", "-1.182", ", ", "-0.351", 
#>     "]", ", ", italic("n")["effects"], " = ", 16L)

# ----------------------- random -----------------------------------------

# creating expression
expr_meta_random(
  data = df,
  type = "random",
  random = "normal",
  output = "dataframe"
)

#> # A tibble: 1 x 8
#>   term    estimate std.error conf.low conf.high statistic  p.value weight
#>   <chr>      <dbl>     <dbl>    <dbl>     <dbl>     <dbl>    <dbl>  <dbl>
#> 1 Overall   -0.746     0.234    -1.26    -0.343     -3.20 0.000501     NA

# ----------------------- Bayes Factor -----------------------------------

# making subtitle
expr_meta_random(
  data = df,
  type = "bayes",
  k = 3,
  # additional arguments given to `metaBMA`
  metaBMA.args = list(
    iter = 5000,
    summarize = "integrate",
    control = list(adapt_delta = 0.99, max_treedepth = 15)
  )
)

#> Warning: Could not get model data.

#> paste("log"["e"], "(BF"["01"], ") = ", "-3.669", ", ", widehat(italic(delta))["mean"]^"posterior", 
#>     " = ", "-0.471", ", CI"["95%"]^"HDI", " [", "-0.803", ", ", 
#>     "-0.155", "]", ", ", italic("r")["Cauchy"]^"JZS", " = ", 
#>     "0.300")

# }