Making expression with frequentist random-effects robust meta-analysis results

This analysis is carried out using the metaplus package. For more, see ?metaplus::metaplus.

expr_meta_robust(data, random = "mixture", k = 2L, messages = TRUE, ...)

Arguments

data
A dataframe. It must contain columns named estimate (effect sizes or outcomes) and std.error (corresponding standard errors). These two columns will be used for yi and sei arguments in metafor::rma (for parametric analysis) or metaplus::metaplus (for robust analysis).

random
The type of random effects distribution. One of "normal","t-dist","mixture", for standard normal, \(t\)-distribution or mixture of normals respectively.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

messages
Deprecated. Retained only for backward compatibility.

...
Arguments passed on to metaplus::metaplus

mods

data frame of covariates corresponding to each study

label

The label to be used for this model when plotting

plotci

Should profile be plotted for each confidence interval?

justfit

Should model only be fitted? If justfit=TRUE then profiling and likelihood ratio statistics are not calculated. Useful for when bootstrapping.

slab

Vector of character strings corresponding to each study.

useAGQ

Use adaptive Gauss-Hermite quadrature when performing integration for t-distribution random-effects. Useful when standard errors for studies are relatively small.

quadpoints

Number of quadrature points to use in adaptive Gauss-Hermite quadrature.

Examples

# \donttest{
# setup
set.seed(123)
library(metaplus)

# renaming to what `statsExpressions` expects
df <- dplyr::rename(mag, estimate = yi, std.error = sei)

# creating expression (changing few defaults)
expr_meta_robust(
  data = df,
  random = "normal",
  k = 4,
  messages = TRUE
)

#> paste("Summary effect: ", italic("z"), " = ", "NA", ", ", italic("p"), 
#>     " = ", "5e-04", ", ", widehat(beta), " = ", "-0.7463", ", CI"["95%"], 
#>     " [", "-1.2583", ", ", "-0.3428", "]", ", ", italic("n")["effects"], 
#>     " = ", 16L)

# }