Expression and dataframe for one-way ANOVA

Expression and dataframe for one-way ANOVA

expr_oneway_anova(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  k = 2L,
  conf.level = 0.95,
  effsize.type = "omega",
  var.equal = FALSE,
  bf.prior = 0.707,
  tr = 0.1,
  nboot = 100,
  top.text = NULL,
  output = "expression",
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
In case of repeated measures design (paired = TRUE, i.e.), this argument specifies the subject or repeated measures id. Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

effsize.type
Type of effect size needed for parametric tests. The argument can be "eta" (partial eta-squared) or "omega" (partial omega-squared).

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

tr
Trim level for the mean when carrying out robust tests. If you get error stating "Standard error cannot be computed because of Winsorized variance of 0 (e.g., due to ties). Try to decrease the trimming level.", try to play around with the value of tr, which is by default set to 0.1. Lowering the value might help.

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100).

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

...
Additional arguments (currently ignored).

Value

For more details, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)

# ----------------------- parametric -------------------------------------

# between-subjects
expr_oneway_anova(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem
)

#> For one-way between subjects designs, partial omega squared is equvilant to omega squared.
#> Returning omega squared.

#> paste(italic("F")["Welch"], "(", "3", ",", "11.10", ") = ", "2.63", 
#>     ", ", italic("p"), " = ", "0.102", ", ", widehat(omega["p"]^2), 
#>     " = ", "0.24", ", CI"["95%"], " [", "0.00", ", ", "0.54", 
#>     "]", ", ", italic("n")["obs"], " = ", "56")

# within-subjects design
expr_oneway_anova(
  data = iris_long,
  x = condition,
  y = value,
  subject.id = id,
  paired = TRUE,
  output = "dataframe"
)

#> # A tibble: 1 x 14
#>   term      sumsq sum.squares.error    df df.error meansq statistic  p.value
#>   <chr>     <dbl>             <dbl> <dbl>    <dbl>  <dbl>     <dbl>    <dbl>
#> 1 condition 1656.              318.  1.15     171.   1.86      776. 1.32e-69
#>   method                                              estimate ci.width conf.low
#>   <chr>                                                  <dbl>    <dbl>    <dbl>
#> 1 ANOVA estimation for factorial designs using 'afex'    0.707     0.95    0.666
#>   conf.high effectsize      
#>       <dbl> <chr>           
#> 1     0.740 Omega2 (partial)

# ----------------------- non-parametric ----------------------------------

# between-subjects
expr_oneway_anova(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem,
  type = "np"
)

#> paste(chi["Kruskal-Wallis"]^2, "(", "3", ") = ", "9.06", ", ", 
#>     italic("p"), " = ", "0.028", ", ", widehat(epsilon)["ordinal"]^2, 
#>     " = ", "0.16", ", CI"["95%"], " [", "0.05", ", ", "0.36", 
#>     "]", ", ", italic("n")["obs"], " = ", "56")

# within-subjects design
expr_oneway_anova(
  data = iris_long,
  x = condition,
  y = value,
  subject.id = id,
  paired = TRUE,
  type = "np",
  output = "dataframe"
)

#> # A tibble: 1 x 9
#>   statistic    df  p.value method                 estimate ci.width conf.low
#>       <dbl> <dbl>    <dbl> <chr>                     <dbl>    <dbl>    <dbl>
#> 1       410     3 1.51e-88 Friedman rank sum test    0.484     0.95    0.340
#>   conf.high effectsize 
#>       <dbl> <chr>      
#> 1     0.969 Kendall's W

# ----------------------- robust -------------------------------------

# between-subjects
expr_oneway_anova(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem,
  type = "r"
)

#> paste(italic("F")["trimmed-means"], "(", "3", ",", "10.57", ") = ", 
#>     "2.51", ", ", italic("p"), " = ", "0.115", ", ", widehat(xi), 
#>     " = ", "0.67", ", CI"["95%"], " [", "0.43", ", ", "0.99", 
#>     "]", ", ", italic("n")["obs"], " = ", "56")

# within-subjects design
expr_oneway_anova(
  data = iris_long,
  x = condition,
  y = value,
  subject.id = id,
  paired = TRUE,
  type = "r",
  output = "dataframe"
)

#> # A tibble: 1 x 10
#>   statistic    df df.error p.value
#>       <dbl> <dbl>    <dbl>   <dbl>
#> 1      559.  1.12     133.       0
#>   method                                                              estimate
#>   <chr>                                                                  <dbl>
#> 1 A heteroscedastic one-way repeated measures ANOVA for trimmed means   -0.204
#>   conf.low conf.high ci.width
#>      <dbl>     <dbl>    <dbl>
#> 1   -0.535     0.173     0.95
#>   effectsize                                                     
#>   <chr>                                                          
#> 1 Algina-Keselman-Penfield robust standardized difference average

# ----------------------- Bayesian -------------------------------------

# between-subjects
expr_oneway_anova(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem,
  type = "bayes"
)

#> paste("log"["e"] * "(BF"["01"] * ") = " * "-1.23" * ", ", widehat(italic(R^"2"))["median"]^"posterior" * 
#>     " = " * "0.11" * ", ", "CI"["95%"]^"HDI" * " [" * "0.00" * 
#>     ", " * "0.27" * "], ", italic("r")["Cauchy"]^"JZS" * " = " * 
#>     "0.71")

# within-subjects design
# needs `BayesFactor 0.9.12-4.3` or above
if (utils::packageVersion("BayesFactor") >= package_version("0.9.12-4.3")) {
  expr_oneway_anova(
    data = iris_long,
    x = condition,
    y = value,
    subject.id = id,
    paired = TRUE,
    type = "bayes",
    output = "dataframe"
  )
}

#> # A tibble: 158 x 19
#>    term                   estimate ci.width conf.low conf.high    pd
#>    <chr>                     <dbl>    <dbl>    <dbl>     <dbl> <dbl>
#>  1 mu                        3.46      0.95    3.35     3.60   1    
#>  2 condition-Petal.Length    0.293     0.95    0.176    0.405  1    
#>  3 condition-Petal.Width    -2.26      0.95   -2.38    -2.15   1    
#>  4 condition-Sepal.Length    2.37      0.95    2.26     2.48   1    
#>  5 condition-Sepal.Width    -0.405     0.95   -0.520   -0.287  1    
#>  6 rowid-1                  -0.662     0.95   -1.39     0.0417 0.962
#>  7 rowid-2                  -0.775     0.95   -1.45    -0.101  0.986
#>  8 rowid-3                  -0.810     0.95   -1.49    -0.0564 0.986
#>  9 rowid-4                  -0.791     0.95   -1.47    -0.0461 0.985
#> 10 rowid-5                  -0.655     0.95   -1.38     0.0499 0.966
#>    rope.percentage prior.distribution prior.location prior.scale effect
#>              <dbl> <chr>                       <dbl>       <dbl> <chr> 
#>  1          0      cauchy                          0       0.707 fixed 
#>  2          0      cauchy                          0       0.707 fixed 
#>  3          0      cauchy                          0       0.707 fixed 
#>  4          0      cauchy                          0       0.707 fixed 
#>  5          0      cauchy                          0       0.707 fixed 
#>  6          0.0424 cauchy                          0       1     random
#>  7          0      cauchy                          0       1     random
#>  8          0      cauchy                          0       1     random
#>  9          0.0115 cauchy                          0       1     random
#> 10          0.0590 cauchy                          0       1     random
#>    component       bf10 method                             r2 std.dev
#>    <chr>          <dbl> <chr>                           <dbl>   <dbl>
#>  1 extra       1.58e182 Bayes factors for linear models 0.818 0.00862
#>  2 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  3 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  4 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  5 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  6 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  7 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  8 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>  9 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#> 10 conditional 1.58e182 Bayes factors for linear models 0.818 0.00862
#>    r2.conf.low r2.conf.high r2.component
#>          <dbl>        <dbl> <chr>       
#>  1       0.800        0.833 conditional 
#>  2       0.800        0.833 conditional 
#>  3       0.800        0.833 conditional 
#>  4       0.800        0.833 conditional 
#>  5       0.800        0.833 conditional 
#>  6       0.800        0.833 conditional 
#>  7       0.800        0.833 conditional 
#>  8       0.800        0.833 conditional 
#>  9       0.800        0.833 conditional 
#> 10       0.800        0.833 conditional 
#> # ... with 148 more rows

# }