Making expression containing t-test results

Making expression containing t-test results

expr_t_twosample(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  k = 2L,
  conf.level = 0.95,
  effsize.type = "g",
  var.equal = FALSE,
  bf.prior = 0.707,
  tr = 0.1,
  nboot = 100,
  output = "expression",
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
A numeric variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
In case of repeated measures design (paired = TRUE, i.e.), this argument specifies the subject or repeated measures id. Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

effsize.type
Type of effect size needed for parametric tests. The argument can be "d" (for Cohen's d) or "g" (for Hedge's g).

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

tr
Trim level for the mean when carrying out robust tests. If you get error stating "Standard error cannot be computed because of Winsorized variance of 0 (e.g., due to ties). Try to decrease the trimming level.", try to play around with the value of tr, which is by default set to 0.1. Lowering the value might help.

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100).

output
If "expression", will return expression with statistical details, while "dataframe" will return a dataframe containing the results.

...
Additional arguments passed to tidyBF::bf_ttest.

Value

Expression containing details from results of a two-sample test and effect size plus confidence intervals.

Note

The stats::wilcox.test function does not follow the same convention as stats::t.test. The sign of the V test statistic will always be positive since it is the sum of the positive signed ranks. Therefore, V will vary in magnitude but not significance based solely on the order of the grouping variable. Consider manually reordering your factor levels if appropriate as shown in the second example below.

References

For more details, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# for reproducibility
set.seed(123)
library(statsExpressions)

# ----------------------- parametric -------------------------------------

# between-subjects design
expr_t_twosample(
  data = sleep,
  x = group,
  y = extra,
  type = "p"
)

#> paste(italic("t")["Welch"], "(", "17.78", ") = ", "-1.86", ", ", 
#>     italic("p"), " = ", "0.079", ", ", widehat(italic("g"))["Hedge"], 
#>     " = ", "-0.80", ", CI"["95%"], " [", "-1.67", ", ", "0.09", 
#>     "]", ", ", italic("n")["obs"], " = ", "20")

# within-subjects design
expr_t_twosample(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "p",
  output = "dataframe"
)

#> # A tibble: 1 x 11
#>   term  group    statistic df.error  p.value method        estimate ci.width
#>   <chr> <chr>        <dbl>    <dbl>    <dbl> <chr>            <dbl>    <dbl>
#> 1 score modality     -3.96       33 0.000373 Paired t-test   -0.664     0.95
#>   conf.low conf.high effectsize
#>      <dbl>     <dbl> <chr>     
#> 1    -1.04    -0.299 Hedges' g 

# ----------------------- non-parametric ----------------------------------

# between-subjects design
expr_t_twosample(
  data = sleep,
  x = group,
  y = extra,
  type = "np"
)

#> paste("log"["e"](italic("W")["Mann-Whitney"]), " = ", "3.24", 
#>     ", ", italic("p"), " = ", "0.069", ", ", widehat(italic("r"))["biserial"]^"rank", 
#>     " = ", "-0.49", ", CI"["95%"], " [", "-0.85", ", ", "-0.07", 
#>     "]", ", ", italic("n")["obs"], " = ", "20")

# within-subjects design
expr_t_twosample(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "np",
  output = "dataframe"
)

#> Warning: Zeros detected. These cannot be sign-rank transformed.

#> # A tibble: 1 x 10
#>   parameter1 parameter2 statistic  p.value method                    estimate
#>   <chr>      <chr>          <dbl>    <dbl> <chr>                        <dbl>
#> 1 score      modality        1.50 0.000886 Wilcoxon signed rank test   -0.934
#>   ci.width conf.low conf.high effectsize       
#>      <dbl>    <dbl>     <dbl> <chr>            
#> 1     0.95       -1    -0.761 r (rank biserial)

# ------------------------------ robust ----------------------------------

# between-subjects design
expr_t_twosample(
  data = sleep,
  x = group,
  y = extra,
  type = "r"
)

#> paste(italic("t")["Yuen"], "(", "13.90", ") = ", "1.57", ", ", 
#>     italic("p"), " = ", "0.139", ", ", widehat(italic(xi)), " = ", 
#>     "0.52", ", CI"["95%"], " [", "0.00", ", ", "0.89", "]", ", ", 
#>     italic("n")["obs"], " = ", "20")

# within-subjects design
expr_t_twosample(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "r",
  output = "dataframe"
)

#> # A tibble: 1 x 12
#>   statistic df.error p.value method                                            
#>       <dbl>    <dbl>   <dbl> <chr>                                             
#> 1     -3.07       27 0.00479 Yuen's test on trimmed means for dependent samples
#>   effectsize `NULL` estimate     S     M     L conf.low conf.high
#>   <chr>       <dbl>    <dbl> <dbl> <dbl> <dbl>    <dbl>     <dbl>
#> 1 AKP             0   -0.593   0.1   0.3   0.5   -0.836    -0.296

#' # ------------------------------ Bayesian ------------------------------

# between-subjects design
expr_t_twosample(
  data = sleep,
  x = group,
  y = extra,
  type = "bayes"
)

#> paste("log"["e"], "(BF"["01"], ") = ", "-0.24", ", ", widehat(italic(delta))["median"]^"posterior", 
#>     " = ", "1.15", ", CI"["95%"]^"HDI", " [", "-0.40", ", ", 
#>     "2.70", "]", ", ", italic("r")["Cauchy"]^"JZS", " = ", "0.71")

# within-subjects design
expr_t_twosample(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "bayes",
  output = "dataframe"
)

#> # A tibble: 2 x 11
#>   term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    0.172   0.0818     0.266 1                   0
#> 2 Cohens_d     -0.637  -0.992     -0.258 0.999               0
#>   prior.distribution prior.location prior.scale  bf10 log_e_bf10
#>   <chr>                       <dbl>       <dbl> <dbl>      <dbl>
#> 1 cauchy                          0       0.707  77.1       4.34
#> 2 cauchy                          0       0.707  77.1       4.34