Expression and dataframe for two-sample t-test

A dataframe containing details from results of a two-sample test and effect size plus confidence intervals.

For more details, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

two_sample_test(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  k = 2L,
  conf.level = 0.95,
  effsize.type = "g",
  var.equal = FALSE,
  bf.prior = 0.707,
  tr = 0.2,
  nboot = 100,
  top.text = NULL,
  ...
)

expr_t_twosample(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  k = 2L,
  conf.level = 0.95,
  effsize.type = "g",
  var.equal = FALSE,
  bf.prior = 0.707,
  tr = 0.2,
  nboot = 100,
  top.text = NULL,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or a table will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
Relevant in case of repeated measures design (paired = TRUE, i.e.), it specifies the subject or repeated measures identifier. Important: Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

type
A character specifying the type of statistical approach. Four possible options:

"parametric"

"nonparametric"

"robust"

"bayes"

Corresponding abbreviations are also accepted: "p" (for parametric), "np" (for nonparametric), "r" (for robust), or "bf" (for Bayesian).

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

effsize.type
Type of effect size needed for parametric tests. The argument can be "d" (for Cohen's d) or "g" (for Hedge's g).

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates.

tr
Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100).

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Currently ignored.

Note

The stats::wilcox.test function does not follow the same convention as stats::t.test. The sign of the V test statistic will always be positive since it is the sum of the positive signed ranks. Therefore, V will vary in magnitude but not significance based solely on the order of the grouping variable. Consider manually reordering your factor levels if appropriate as shown in the second example below.

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)

# ----------------------- parametric -------------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x = group,
  y = extra,
  type = "p"
)

#> # A tibble: 1 x 14
#>   term  group mean.group1 mean.group2 statistic df.error p.value method estimate
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl> <chr>     <dbl>
#> 1 extra group        0.75        2.33     -1.86     17.8  0.0794 Welch…   -0.797
#> # … with 5 more variables: conf.level <dbl>, conf.low <dbl>, conf.high <dbl>,
#> #   effectsize <chr>, expression <list>

# within-subjects design
two_sample_test(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "p"
)

#> # A tibble: 1 x 12
#>   term  group   statistic df.error p.value method   estimate conf.level conf.low
#>   <chr> <chr>       <dbl>    <dbl>   <dbl> <chr>       <dbl>      <dbl>    <dbl>
#> 1 score modali…     -3.96       33 3.73e-4 Paired …   -0.664       0.95    -1.04
#> # … with 3 more variables: conf.high <dbl>, effectsize <chr>, expression <list>

# ----------------------- non-parametric ----------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x = group,
  y = extra,
  type = "np"
)

#> # A tibble: 1 x 11
#>   parameter1 parameter2 statistic p.value method    estimate conf.level conf.low
#>   <chr>      <chr>          <dbl>   <dbl> <chr>        <dbl>      <dbl>    <dbl>
#> 1 extra      group           3.24  0.0693 Wilcoxon…    -0.49       0.95   -0.850
#> # … with 3 more variables: conf.high <dbl>, effectsize <chr>, expression <list>

# within-subjects design
two_sample_test(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "np"
)

#> # A tibble: 1 x 11
#>   parameter1 parameter2 statistic  p.value method   estimate conf.level conf.low
#>   <chr>      <chr>          <dbl>    <dbl> <chr>       <dbl>      <dbl>    <dbl>
#> 1 score      modality        1.50 0.000886 Wilcoxo…   -0.934       0.95       -1
#> # … with 3 more variables: conf.high <dbl>, effectsize <chr>, expression <list>

# ------------------------------ robust ----------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x = group,
  y = extra,
  type = "r"
)

#> # A tibble: 1 x 10
#>   statistic df.error p.value method       estimate conf.low conf.high conf.level
#>       <dbl>    <dbl>   <dbl> <chr>           <dbl>    <dbl>     <dbl>      <dbl>
#> 1      1.62     8.26   0.143 Yuen's test…    0.516        0     0.877       0.95
#> # … with 2 more variables: effectsize <chr>, expression <list>

# within-subjects design
two_sample_test(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "r"
)

#> Warning: the standard deviation is zero

#> # A tibble: 1 x 10
#>   statistic df.error p.value method       estimate conf.low conf.high conf.level
#>       <dbl>    <dbl>   <dbl> <chr>           <dbl>    <dbl>     <dbl>      <dbl>
#> 1     -2.56       21  0.0182 Yuen's test…   -0.381   -0.704    -0.238       0.95
#> # … with 2 more variables: effectsize <chr>, expression <list>

#' # ------------------------------ Bayesian ------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x = group,
  y = extra,
  type = "bayes"
)

#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    1.15        0.95   -0.400     2.70  0.934          0.0674
#> 2 Cohens_d     -0.601       0.95   -1.50      0.210 0.930          0.0834
#> # … with 6 more variables: prior.distribution <chr>, prior.location <dbl>,
#> #   prior.scale <dbl>, bf10 <dbl>, method <chr>, expression <list>

# within-subjects design
two_sample_test(
  data = VR_dilemma,
  x = modality,
  y = score,
  paired = TRUE,
  subject.id = id,
  type = "bayes"
)

#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference    0.172       0.95   0.0818     0.266 1                   0
#> 2 Cohens_d     -0.637       0.95  -0.992     -0.258 0.999               0
#> # … with 6 more variables: prior.distribution <chr>, prior.location <dbl>,
#> #   prior.scale <dbl>, bf10 <dbl>, method <chr>, expression <list>

# }