Dataframe and expression for distribution properties

Dataframe and expression for distribution properties

Usage

centrality_description(data, x, y, type = "parametric", tr = 0.2, k = 2L, ...)

Arguments

data

A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x

The grouping (or independent) variable from the dataframe data.

y

The response (or outcome or dependent) variable from the dataframe data.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

...

Currently ignored.

Details

This function describes a distribution for y variable for each level of the grouping variable in x by a set of indices (e.g., measures of centrality, dispersion, range, skewness, kurtosis). It additionally returns an expression containing a specified centrality measure. The function internally relies on datawizard::describe_distribution function.

Examples

set.seed(123)

# parametric -----------------------
centrality_description(iris, Species, Sepal.Length)
#> # A tibble: 3 × 15
#>   Species    Sepal.Length n_obs variable  std.dev   iqr conf.low conf.high   min
#>   <fct>             <dbl> <int> <chr>       <dbl> <dbl>    <dbl>     <dbl> <dbl>
#> 1 setosa             5.01    50 Sepal.Le…   0.352 0.400     4.90      5.10   4.3
#> 2 versicolor         5.94    50 Sepal.Le…   0.516 0.7       5.80      6.07   4.9
#> 3 virginica          6.59    50 Sepal.Le…   0.636 0.750     6.39      6.79   4.9
#> # … with 6 more variables: max <dbl>, skewness <dbl>, kurtosis <dbl>,
#> #   n.missing <int>, expression <glue>, n_label <chr>

# non-parametric -------------------
centrality_description(mtcars, am, wt, type = "n")
#> # A tibble: 2 × 15
#>   am       wt n_obs variable   mad   iqr conf.low conf.high   min   max skewness
#>   <fct> <dbl> <int> <chr>    <dbl> <dbl>    <dbl>     <dbl> <dbl> <dbl>    <dbl>
#> 1 0      3.52    19 wt       0.452 0.41      3.44      3.84  2.46  5.42    1.15 
#> 2 1      2.32    13 wt       0.682 0.942     1.94      2.78  1.51  3.57    0.269
#> # … with 4 more variables: kurtosis <dbl>, n.missing <int>, expression <glue>,
#> #   n_label <chr>

# robust ---------------------------
centrality_description(ToothGrowth, supp, len, type = "r")
#> # A tibble: 2 × 15
#>   supp    len n_obs variable std.dev   iqr conf.low conf.high   min   max
#>   <fct> <dbl> <int> <chr>      <dbl> <dbl>    <dbl>     <dbl> <dbl> <dbl>
#> 1 OJ     21.7    30 len         6.61  10.9     17.5      23.5   8.2  30.9
#> 2 VC     16.6    30 len         8.27  12.5     13.6      19.6   4.2  33.9
#> # … with 5 more variables: skewness <dbl>, kurtosis <dbl>, n.missing <int>,
#> #   expression <glue>, n_label <chr>

# Bayesian -------------------------
centrality_description(sleep, group, extra, type = "b")
#> # A tibble: 2 × 14
#>   group  extra n_obs variable   iqr conf.low conf.high   min   max skewness
#>   <fct>  <dbl> <int> <chr>    <dbl>    <dbl>     <dbl> <dbl> <dbl>    <dbl>
#> 1 1     0.0579    10 extra     2.8   -1.39        3.65  -1.6   3.7    0.581
#> 2 2     0.973     10 extra     3.82   0.0728      4.88  -0.1   5.5    0.386
#> # … with 4 more variables: kurtosis <dbl>, n.missing <int>, expression <glue>,
#> #   n_label <chr>