Contingency table analyses

A dataframe containing results from for contingency table analysis or goodness of fit test.

To see details about functions which are internally used to carry out these analyses, see the following vignette- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

contingency_table(
  data,
  x,
  y = NULL,
  paired = FALSE,
  type = "parametric",
  counts = NULL,
  ratio = NULL,
  k = 2L,
  conf.level = 0.95,
  sampling.plan = "indepMulti",
  fixed.margin = "rows",
  prior.concentration = 1,
  top.text = NULL,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x
The variable to use as the rows in the contingency table.

y
The variable to use as the columns in the contingency table. Default is NULL. If NULL, one-sample proportion test (a goodness of fit test) will be run for the x variable. Otherwise association test will be carried out.

paired
Logical indicating whether data came from a within-subjects or repeated measures design study (Default: FALSE). If TRUE, McNemar's test expression will be returned. If FALSE, Pearson's chi-square test will be returned.

type
A character specifying the type of statistical approach.

"parametric"

"nonparametric"

"robust"

"bayes"

Corresponding abbreviations are also accepted: "p" (for parametric), "np" (for nonparametric), "r" (for robust), or "bf" (for Bayesian).

counts
A string naming a variable in data containing counts, or NULL if each row represents a single observation.

ratio
A vector of proportions: the expected proportions for the proportion test (should sum to 1). Default is NULL, which means the null is equal theoretical proportions across the levels of the nominal variable. This means if there are two levels this will be ratio = c(0.5,0.5) or if there are four levels this will be ratio = c(0.25,0.25,0.25,0.25), etc.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

sampling.plan
Character describing the sampling plan. Possible options are "indepMulti" (independent multinomial; default), "poisson", "jointMulti" (joint multinomial), "hypergeom" (hypergeometric). For more, see ?BayesFactor::contingencyTableBF().

fixed.margin
For the independent multinomial sampling plan, which margin is fixed ("rows" or "cols"). Defaults to "rows".

prior.concentration
Specifies the prior concentration parameter, set to 1 by default. It indexes the expected deviation from the null hypothesis under the alternative, and corresponds to Gunel and Dickey's (1974) "a" parameter.

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Additional arguments (currently ignored).

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# ------------------------ non-Bayesian -----------------------------

# association test
contingency_table(
  data = mtcars,
  x = am,
  y = cyl,
  paired = FALSE
)

#> Warning: Chi-squared approximation may be incorrect

#> # A tibble: 1 x 10
#>   statistic    df p.value method                     estimate conf.level
#>       <dbl> <int>   <dbl> <chr>                         <dbl>      <dbl>
#> 1      8.74     2  0.0126 Pearson's Chi-squared test    0.464       0.95
#>   conf.low conf.high effectsize        expression
#>      <dbl>     <dbl> <chr>             <list>    
#> 1        0     0.777 Cramer's V (adj.) <language>

# goodness-of-fit test
contingency_table(
  data = as.data.frame(HairEyeColor),
  x = Eye,
  counts = Freq,
  ratio = c(0.2, 0.2, 0.3, 0.3)
)

#> # A tibble: 1 x 10
#>   statistic    df  p.value method                                   estimate
#>       <dbl> <dbl>    <dbl> <chr>                                       <dbl>
#> 1      279.     3 3.56e-60 Chi-squared test for given probabilities    0.395
#>   conf.level conf.low conf.high effectsize        expression
#>        <dbl>    <dbl>     <dbl> <chr>             <list>    
#> 1       0.95    0.346     0.439 Cramer's V (adj.) <language>

# ------------------------ Bayesian -----------------------------

# association test
contingency_table(
  data = mtcars,
  x = am,
  y = cyl,
  paired = FALSE,
  type = "bayes"
)

#> Warning: Could not estimate a good default ROPE range. Using 'c(-0.1, 0.1)'.

#> # A tibble: 8 x 13
#>   term      estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>        <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 cell[1,1]    3.15        0.95    0.692     6.55      1               0
#> 2 cell[2,1]    7.47        0.95    3.34     11.7       1               0
#> 3 cell[1,2]    3.96        0.95    1.20      7.70      1               0
#> 4 cell[2,2]    3.18        0.95    0.778     6.51      1               0
#> 5 cell[1,3]   10.8         0.95    6.56     15.9       1               0
#> 6 cell[2,3]    2.27        0.95    0.332     5.31      1               0
#> 7 Ratio       NA          NA      NA        NA        NA              NA
#> 8 Cramers_v    0.478       0.95    0.221     0.725     1               0
#>   prior.distribution      prior.location prior.scale  bf10
#>   <chr>                            <dbl>       <dbl> <dbl>
#> 1 independent multinomial              0           1  16.8
#> 2 independent multinomial              0           1  16.8
#> 3 independent multinomial              0           1  16.8
#> 4 independent multinomial              0           1  16.8
#> 5 independent multinomial              0           1  16.8
#> 6 independent multinomial              0           1  16.8
#> 7 independent multinomial              0           1  16.8
#> 8 independent multinomial              0           1  16.8
#>   method                              expression
#>   <chr>                               <list>    
#> 1 Bayesian contingency table analysis <language>
#> 2 Bayesian contingency table analysis <language>
#> 3 Bayesian contingency table analysis <language>
#> 4 Bayesian contingency table analysis <language>
#> 5 Bayesian contingency table analysis <language>
#> 6 Bayesian contingency table analysis <language>
#> 7 Bayesian contingency table analysis <language>
#> 8 Bayesian contingency table analysis <language>

# goodness-of-fit test
contingency_table(
  data = as.data.frame(HairEyeColor),
  x = Eye,
  counts = Freq,
  ratio = c(0.2, 0.2, 0.3, 0.3),
  type = "bayes"
)

#> # A tibble: 1 x 3
#>      bf10 prior.scale expression
#>     <dbl>       <dbl> <list>    
#> 1 3.74e55           1 <language>

# }