Correlation analyses

A dataframe containing results from correlation test with confidence intervals for the correlation coefficient estimate.

corr_test(
  data,
  x,
  y,
  type = "parametric",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  top.text = NULL,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x
The column in data containing the explanatory variable to be plotted on the x-axis.

y
The column in data containing the response (outcome) variable to be plotted on the y-axis.

type
A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

tr
Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Additional arguments (currently ignored).

References

To see details about functions which are internally used to carry out these analyses, see the following vignette- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# without changing defaults
corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack
)

#> # A tibble: 1 x 13
#>   parameter1 parameter2 estimate conf.level conf.low conf.high statistic
#>   <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>     <dbl>
#> 1 area       percblack   -0.0465       0.95   -0.140    0.0475    -0.970
#>   df.error p.value method              n.obs effectsize          expression
#>      <int>   <dbl> <chr>               <int> <chr>               <list>    
#> 1      435   0.332 Pearson correlation   437 Pearson correlation <language>

# changing defaults
corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack,
  type = "robust"
)

#> # A tibble: 1 x 13
#>   parameter1 parameter2 estimate conf.level conf.low conf.high statistic
#>   <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>     <dbl>
#> 1 area       percblack    0.0293       0.95  -0.0647     0.123     0.611
#>   df.error p.value method                         n.obs
#>      <int>   <dbl> <chr>                          <int>
#> 1      435   0.541 Winsorized Pearson correlation   437
#>   effectsize                     expression
#>   <chr>                          <list>    
#> 1 Winsorized Pearson correlation <language>

# }