Correlation analyses

The table below provides summary about:

statistical test carried out for inferential statistics

type of effect size estimate and a measure of uncertainty for this estimate

functions used internally to compute these details

Hypothesis testing and Effect size estimation

Type	Test	CI available?	Function used

Parametric	Pearson's correlation coefficient	Yes	correlation::correlation()

Non-parametric	Spearman's rank correlation coefficient	Yes	correlation::correlation()

Robust	Winsorized Pearson correlation coefficient	Yes	correlation::correlation()

Bayesian	Bayesian Pearson's correlation coefficient	Yes	correlation::correlation()

Usage

corr_test(
  data,
  x,
  y,
  type = "parametric",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  ...
)

Arguments

data

A data frame (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted. Additionally, grouped data frames from {dplyr} should be ungrouped before they are entered as data.

x

The column in data containing the explanatory variable to be plotted on the x-axis.

y

The column in data containing the response (outcome) variable to be plotted on the y-axis.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level

Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior

A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

...

Additional arguments (currently ignored).

Value

The returned tibble data frame can contain some or all of the following columns (the exact columns will depend on the statistical test):

statistic: the numeric value of a statistic

df: the numeric value of a parameter being modeled (often degrees of freedom for the test)

df.error and df: relevant only if the statistic in question has two degrees of freedom (e.g. anova)

p.value: the two-sided p-value associated with the observed statistic

method: the name of the inferential statistical test

estimate: estimated value of the effect size

conf.low: lower bound for the effect size estimate

conf.high: upper bound for the effect size estimate

conf.level: width of the confidence interval

conf.method: method used to compute confidence interval

conf.distribution: statistical distribution for the effect

effectsize: the name of the effect size

n.obs: number of observations

expression: pre-formatted expression containing statistical details

For examples of dataframe outputs, see examples and this vignette.

Note that all examples are preceded by set.seed() calls for reproducibility.

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# without changing defaults
corr_test(
  data = ggplot2::midwest,
  x    = area,
  y    = percblack
)
#> # A tibble: 1 × 13
#>   parameter1 parameter2 effectsize          estimate conf.level conf.low
#>   <chr>      <chr>      <chr>                  <dbl>      <dbl>    <dbl>
#> 1 area       percblack  Pearson correlation  -0.0465       0.95   -0.140
#>   conf.high statistic df.error p.value method              n.obs expression
#>       <dbl>     <dbl>    <int>   <dbl> <chr>               <int> <list>    
#> 1    0.0475    -0.970      435   0.332 Pearson correlation   437 <language>

# changing defaults
corr_test(
  data = ggplot2::midwest,
  x    = area,
  y    = percblack,
  type = "robust"
)
#> # A tibble: 1 × 13
#>   parameter1 parameter2 effectsize                     estimate conf.level
#>   <chr>      <chr>      <chr>                             <dbl>      <dbl>
#> 1 area       percblack  Winsorized Pearson correlation   0.0293       0.95
#>   conf.low conf.high statistic df.error p.value method                        
#>      <dbl>     <dbl>     <dbl>    <int>   <dbl> <chr>                         
#> 1  -0.0647     0.123     0.611      435   0.541 Winsorized Pearson correlation
#>   n.obs expression
#>   <int> <list>    
#> 1   437 <language>
# }