A dataframe with expression for correlation analysis

A dataframe containing results from correlation test with confidence intervals for the correlation coefficient estimate. Results are extracted via correlation::correlation.

corr_test(
  data,
  x,
  y,
  type = "parametric",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  top.text = NULL,
  ...
)

expr_corr_test(
  data,
  x,
  y,
  type = "parametric",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  top.text = NULL,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x
The column in data containing the explanatory variable to be plotted on the x-axis. Can be entered either as a character string (e.g., "x") or as a bare expression (e.g, x).

y
The column in data containing the response (outcome) variable to be plotted on the y-axis. Can be entered either as a character string (e.g., "y") or as a bare expression (e.g, y).

type
A character specifying the type of statistical approach. Four possible options:

"parametric"

"nonparametric"

"robust"

"bayes"

Corresponding abbreviations are also accepted: "p" (for parametric), "np" (for nonparametric), "r" (for robust), or "bf" (for Bayesian).

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

tr
Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates.

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Additional arguments (currently ignored).

References

For more details, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# for reproducibility
set.seed(123)
library(statsExpressions)

# without changing defaults
corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack
)

#> # A tibble: 1 x 13
#>   parameter1 parameter2 estimate conf.level conf.low conf.high statistic
#>   <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>     <dbl>
#> 1 area       percblack   -0.0465       0.95   -0.140    0.0475    -0.970
#> # … with 6 more variables: df.error <int>, p.value <dbl>, method <chr>,
#> #   n.obs <int>, effectsize <chr>, expression <list>

# changing defaults
corr_test(
  data = ggplot2::midwest,
  x = area,
  y = percblack,
  type = "robust"
)

#> # A tibble: 1 x 13
#>   parameter1 parameter2 estimate conf.level conf.low conf.high statistic
#>   <chr>      <chr>         <dbl>      <dbl>    <dbl>     <dbl>     <dbl>
#> 1 area       percblack    0.0293       0.95  -0.0647     0.123     0.611
#> # … with 6 more variables: df.error <int>, p.value <dbl>, method <chr>,
#> #   n.obs <int>, effectsize <chr>, expression <list>