Random-effects meta-analysis

Parametric, non-parametric, robust, and Bayesian random-effects meta-analysis.

Usage

meta_analysis(
  data,
  type = "parametric",
  random = "mixture",
  k = 2L,
  conf.level = 0.95,
  ...
)

Arguments

data

A data frame. It must contain columns named estimate (effect sizes or outcomes) and std.error (corresponding standard errors). These two columns will be used:

as yi and sei arguments in metafor::rma() (for parametric test) or metaplus::metaplus() (for robust test)

as y and SE arguments in metaBMA::meta_random() (for Bayesian test).

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

random

The type of random effects distribution. One of "normal", "t-dist", "mixture", for standard normal, \(t\)-distribution or mixture of normals respectively.

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level

Scalar between 0 and 1 (default: 95% confidence/credible intervals, 0.95). If NULL, no confidence intervals will be computed.

...

Additional arguments passed to the respective meta-analysis function.

Value

The returned tibble data frame can contain some or all of the following columns (the exact columns will depend on the statistical test):

statistic: the numeric value of a statistic

df: the numeric value of a parameter being modeled (often degrees of freedom for the test)

df.error and df: relevant only if the statistic in question has two degrees of freedom (e.g. anova)

p.value: the two-sided p-value associated with the observed statistic

method: the name of the inferential statistical test

estimate: estimated value of the effect size

conf.low: lower bound for the effect size estimate

conf.high: upper bound for the effect size estimate

conf.level: width of the confidence interval

conf.method: method used to compute confidence interval

conf.distribution: statistical distribution for the effect

effectsize: the name of the effect size

n.obs: number of observations

expression: pre-formatted expression containing statistical details

For examples, see data frame output vignette.

Note

Important: The function assumes that you have already downloaded the needed package ({metafor}, {metaplus}, or {metaBMA}) for meta-analysis. If they are not available, you will be asked to install them.

Random-effects meta-analysis

The table below provides summary about:

statistical test carried out for inferential statistics

type of effect size estimate and a measure of uncertainty for this estimate

functions used internally to compute these details

Hypothesis testing and Effect size estimation

Type	Test	CI available?	Function used

Parametric	Pearson's correlation coefficient	Yes	correlation::correlation()

Non-parametric	Spearman's rank correlation coefficient	Yes	correlation::correlation()

Robust	Winsorized Pearson's correlation coefficient	Yes	correlation::correlation()

Bayesian	Bayesian Pearson's correlation coefficient	Yes	correlation::correlation()

Examples

# setup
set.seed(123)
library(statsExpressions)

# a data frame with estimates and standard errors
# (`mag` dataset from `{metaplus}`)
df <- tibble::tribble(
  ~study, ~estimate, ~std.error,
  "Abraham", -0.83, 1.247,
  "Bertschat", -1.056, 0.414,
  "Ceremuzynski", -1.278, 0.808,
  "Feldstedt", -0.043, 1.429,
  "Golf", 0.223, 0.489,
  "ISIS-4", -2.407, 1.072,
  "LIMIT-2", -1.28, 1.193,
  "Morton", -1.191, 1.661,
  "Pereira", -0.695, 0.536,
  "Rasmussen", -2.208, 1.109,
  "Schechter", -2.038, 0.78,
  "Schechter 1", -0.85, 0.618,
  "Schechter 2", -0.793, 0.625,
  "Singh", -0.299, 0.146,
  "Smith", -1.57, 0.574,
  "Thogersen", 0.057, 0.031
)

# parametric
meta_analysis(df)
#> # A tibble: 1 × 14
#>   term    effectsize                     estimate std.error conf.level conf.low
#>   <chr>   <chr>                             <dbl>     <dbl>      <dbl>    <dbl>
#> 1 Overall meta-analytic summary estimate   -0.766     0.212       0.95    -1.18
#>   conf.high statistic  p.value weight method                        conf.method
#>       <dbl>     <dbl>    <dbl>  <dbl> <chr>                         <chr>      
#> 1    -0.351     -3.62 0.000295     NA Meta-analysis using 'metafor' Wald       
#>   n.obs expression
#>   <int> <list>    
#> 1    16 <language>
# robust
meta_analysis(df, type = "random", random = "normal")
#> # A tibble: 1 × 14
#>   term    effectsize                     estimate std.error conf.low conf.high
#>   <chr>   <chr>                             <dbl>     <dbl>    <dbl>     <dbl>
#> 1 Overall meta-analytic summary estimate   -0.746     0.233    -1.26    -0.343
#>   statistic  p.value weight conf.level method                               
#>       <dbl>    <dbl>  <dbl>      <dbl> <chr>                                
#> 1     -3.20 0.000500     NA       0.95 Robust meta-analysis using 'metaplus'
#>   conf.method n.obs expression
#>   <chr>       <int> <list>    
#> 1 Wald           16 <language>
# Bayesian
meta_analysis(df, type = "bayes")
#> Warning: There were 1 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> # A tibble: 2 × 20
#>   term    effectsize                       estimate std.error conf.level
#>   <chr>   <chr>                               <dbl>     <dbl>      <dbl>
#> 1 Overall meta-analytic posterior estimate   -0.647     0.219       0.95
#> 2 tau     meta-analytic posterior estimate    0.489     0.184       0.95
#>   conf.low conf.high weight  bf10  rhat   ess component prior.distribution
#>      <dbl>     <dbl>  <dbl> <dbl> <dbl> <dbl> <chr>     <chr>             
#> 1   -1.11     -0.250     NA  53.1     1 5064. meta      Student's t       
#> 2    0.208     0.924     NA  53.1     1 4093. meta      Inverse gamma     
#>   prior.location prior.scale method                                 conf.method
#>            <dbl>       <dbl> <chr>                                  <chr>      
#> 1              0       0.707 Bayesian meta-analysis using 'metaBMA' ETI        
#> 2              1       0.15  Bayesian meta-analysis using 'metaBMA' ETI        
#>   log_e_bf10 n.obs expression
#>        <dbl> <int> <list>    
#> 1       3.97    16 <language>
#> 2       3.97    16 <language>