Random-effects meta-analyses

A dataframe containing results from random-effects meta-analysis.

To see details about functions which are internally used to carry out these analyses, see the following vignette- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

meta_analysis(
  data,
  type = "parametric",
  random = "mixture",
  k = 2L,
  conf.level = 0.95,
  top.text = NULL,
  ...
)

Arguments

data
A dataframe. It must contain columns named estimate (effect sizes or outcomes) and std.error (corresponding standard errors). These two columns will be used for yi and sei arguments in metafor::rma (for parametric analysis) or metaplus::metaplus (for robust analysis), and for y and SE arguments in metaBMA::meta_random (for Bayesian analysis).

type
A character specifying the type of statistical approach. Four possible options:

"parametric"

"nonparametric"

"robust"

"bayes"

Corresponding abbreviations are also accepted: "p" (for parametric), "np" (for nonparametric), "r" (for robust), or "bf" (for Bayesian).

random
The type of random effects distribution. One of "normal","t-dist","mixture", for standard normal, \(t\)-distribution or mixture of normals respectively.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Additional arguments passed to the respective meta-analysis function.

Note

Important: The function assumes that you have already downloaded the needed package (metafor, metaplus, or metaBMA) for meta-analysis. If they are not available, you will be asked to install them.

Examples

# \donttest{
# run examples only if the needed packages are available
if (all(unlist(lapply(
  c("metaplus", "metafor", "metaBMA"), # needed packages
  require,
  character.only = TRUE,
  quietly = TRUE,
  warn.conflicts = FALSE
)))) {
  # note that the `print` calls below are not necessary for you to write
  # they are in the documentation so that the website renders them

  # setup
  set.seed(123)
  library(statsExpressions)
  options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

  # renaming to what `statsExpressions` expects
  df <- dplyr::rename(mag, estimate = yi, std.error = sei)

  # ----------------------- parametric ---------------------------------------

  print(meta_analysis(data = df))

  # ----------------------- random -----------------------------------------

  print(meta_analysis(
    data = df,
    type = "random",
    random = "normal"
  ))

  # ----------------------- Bayes Factor -----------------------------------

  meta_analysis(
    data = df,
    type = "bayes",

    # additional arguments given to `metaBMA`
    iter = 5000,
    summarize = "integrate",
    control = list(adapt_delta = 0.99, max_treedepth = 15)
  )
}

#> 
#> Note: The second parameter is the study standard error not the standard error squared as in the metafor package.

#> Loading 'metafor' package (version 2.4-0). For an overview 
#> and introduction to the package please type: help(metafor).

#> This is metaBMA version 0.6.7

#> - Default priors were changed in version 0.6.6.

#> - Since default priors may change again, it is safest to specify priors (even when using the defaults).

#> # A tibble: 1 x 12
#>   term    estimate std.error conf.level conf.low conf.high statistic  p.value
#>   <chr>      <dbl>     <dbl>      <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 Overall   -0.767     0.212       0.95    -1.18    -0.351     -3.62 0.000295
#>   weight method                        effectsize                     expression
#>    <dbl> <chr>                         <chr>                          <list>    
#> 1     NA Meta-analysis using 'metafor' meta-analytic summary estimate <language>
#> # A tibble: 1 x 12
#>   term    estimate std.error conf.low conf.high statistic  p.value weight
#>   <chr>      <dbl>     <dbl>    <dbl>     <dbl>     <dbl>    <dbl>  <dbl>
#> 1 Overall   -0.746     0.234    -1.26    -0.343     -3.20 0.000501     NA
#>   conf.level method                               
#>        <dbl> <chr>                                
#> 1       0.95 Robust meta-analysis using 'metaplus'
#>   effectsize                     expression
#>   <chr>                          <list>    
#> 1 meta-analytic summary estimate <language>

#> # A tibble: 2 x 17
#>   term    estimate std.error conf.level conf.low conf.high weight  bf10  rhat
#>   <chr>      <dbl>     <dbl>      <dbl>    <dbl>     <dbl>  <dbl> <dbl> <dbl>
#> 1 Overall   -0.650     0.222       0.95   -1.10     -0.235     NA  53.0    NA
#> 2 tau        0.486     0.184       0.95    0.167     0.848     NA  53.0    NA
#>     ess component prior.distribution prior.location prior.scale
#>   <dbl> <chr>     <chr>                       <dbl>       <dbl>
#> 1    NA meta      Student's t                     0       0.707
#> 2    NA meta      Inverse gamma                   1       0.15 
#>   method                                 effectsize                      
#>   <chr>                                  <chr>                           
#> 1 Bayesian meta-analysis using 'metaBMA' meta-analytic posterior estimate
#> 2 Bayesian meta-analysis using 'metaBMA' meta-analytic posterior estimate
#>   expression
#>   <list>    
#> 1 <language>
#> 2 <language>

# }