One-sample tests

The table below provides summary about:

statistical test carried out for inferential statistics

type of effect size estimate and a measure of uncertainty for this estimate

functions used internally to compute these details

Hypothesis testing

Type	Test	Function used

Parametric	One-sample Student's t-test	stats::t.test

Non-parametric	One-sample Wilcoxon test	stats::wilcox.test

Robust	Bootstrap-t method for one-sample test	WRS2::trimcibt

Bayesian	One-sample Student's t-test	BayesFactor::ttestBF

Effect size estimation

Type	Effect size	CI available?	Function used

Parametric	Cohen's d, Hedge's g	Yes	effectsize::cohens_d(), effectsize::hedges_g()

Non-parametric	r (rank-biserial correlation)	Yes	effectsize::rank_biserial()

Robust	trimmed mean	Yes	WRS2::trimcibt()

Bayes Factor	difference	Yes	bayestestR::describe_posterior()

Usage

one_sample_test(
  data,
  x,
  type = "parametric",
  test.value = 0,
  alternative = "two.sided",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  effsize.type = "g",
  top.text = NULL,
  ...
)

Arguments

data

A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x

A numeric variable from the dataframe data.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

test.value

A number indicating the true value of the mean (Default: 0).

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level

Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior

A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

effsize.type

Type of effect size needed for parametric tests. The argument can be "d" (for Cohen's d) or "g" (for Hedge's g).

top.text

Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot package functions.

...

Currently ignored.

Value

The returned tibble dataframe can contain some or all of the following columns (the exact columns will depend on the statistical test):

statistic: the numeric value of a statistic

df: the numeric value of a parameter being modeled (often degrees of freedom for the test)

df.error and df: relevant only if the statistic in question has two degrees of freedom (e.g. anova)

p.value: the two-sided p-value associated with the observed statistic

method: the name of the inferential statistical test

estimate: estimated value of the effect size

conf.low: lower bound for the effect size estimate

conf.high: upper bound for the effect size estimate

conf.level: width of the confidence interval

conf.method: method used to compute confidence interval

conf.distribution: statistical distribution for the effect

effectsize: the name of the effect size

n.obs: number of observations

expression: pre-formatted expression containing statistical details

For examples of dataframe outputs, see examples and this vignette. Note that all examples are preceded by set.seed() calls for reproducibility.

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# ----------------------- parametric ---------------------------------------

one_sample_test(
  data       = ggplot2::msleep,
  x          = brainwt,
  test.value = 0.275,
  type       = "parametric"
)
#> # A tibble: 1 × 15
#>      mu statistic df.error p.value method            alternative effectsize
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>             <chr>       <chr>     
#> 1 0.275    0.0504       55   0.960 One Sample t-test two.sided   Hedges' g 
#>   estimate conf.level conf.low conf.high conf.method conf.distribution n.obs
#>      <dbl>      <dbl>    <dbl>     <dbl> <chr>       <chr>             <int>
#> 1  0.00665       0.95   -0.254     0.267 ncp         t                    56
#>   expression  
#>   <list>      
#> 1 <expression>

# ----------------------- non-parametric -----------------------------------

one_sample_test(
  data       = ggplot2::msleep,
  x          = brainwt,
  test.value = 0.275,
  type       = "nonparametric"
)
#> # A tibble: 1 × 12
#>   statistic    p.value method                    alternative effectsize       
#>       <dbl>      <dbl> <chr>                     <chr>       <chr>            
#> 1       248 0.00000738 Wilcoxon signed rank test two.sided   r (rank biserial)
#>   estimate conf.level conf.low conf.high conf.method n.obs expression  
#>      <dbl>      <dbl>    <dbl>     <dbl> <chr>       <int> <list>      
#> 1   -0.689       0.95   -0.817    -0.497 normal         56 <expression>

# ----------------------- robust --------------------------------------------

one_sample_test(
  data       = ggplot2::msleep,
  x          = brainwt,
  test.value = 0.275,
  type       = "robust"
)
#> # A tibble: 1 × 10
#>   statistic p.value n.obs method                                 effectsize  
#>       <dbl>   <dbl> <int> <chr>                                  <chr>       
#> 1     -14.7       0    56 Bootstrap-t method for one-sample test Trimmed mean
#>   estimate conf.level conf.low conf.high expression  
#>      <dbl>      <dbl>    <dbl>     <dbl> <list>      
#> 1   0.0390       0.95  -0.0132    0.0911 <expression>

# ---------------------------- Bayesian -----------------------------------

one_sample_test(
  data       = ggplot2::msleep,
  x          = brainwt,
  test.value = 0.275,
  type       = "bayes",
  bf.prior   = 0.8
)
#> # A tibble: 1 × 16
#>   term       effectsize      estimate conf.level conf.low conf.high    pd
#>   <chr>      <chr>              <dbl>      <dbl>    <dbl>     <dbl> <dbl>
#> 1 Difference Bayesian t-test -0.00925       0.95   -0.263     0.246 0.532
#>   rope.percentage prior.distribution prior.location prior.scale  bf10
#>             <dbl> <chr>                       <dbl>       <dbl> <dbl>
#> 1           0.582 cauchy                          0         0.8 0.130
#>   method          log_e_bf10 n.obs expression  
#>   <chr>                <dbl> <int> <list>      
#> 1 Bayesian t-test      -2.04    56 <expression>
# }