One-sample tests

A dataframe containing results from a one-sample test.

one_sample_test(
  data,
  x,
  type = "parametric",
  test.value = 0,
  alternative = "two.sided",
  k = 2L,
  conf.level = 0.95,
  tr = 0.2,
  bf.prior = 0.707,
  effsize.type = "g",
  top.text = NULL,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x
A numeric variable from the dataframe data.

type
A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

test.value
A number indicating the true value of the mean (Default: 0).

alternative
a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

tr
Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

effsize.type
Type of effect size needed for parametric tests. The argument can be "d" (for Cohen's d) or "g" (for Hedge's g).

top.text
Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot functions.

...
Currently ignored.

Details

The exact test and the effect size details contained will depend on the type argument.

Internal function .f used to carry out statistical test:

parametric: stats::t.test

nonparametric: stats::wilcox.test

robust: trimcibt (custom)

bayes: BayesFactor::ttestBF

Internal function .f.es used to compute effect size:

parametric: effectsize::cohens_d, effectsize::hedges_g

nonparametric: effectsize::rank_biserial

robust: trimcibt (custom)

bayes: bayestestR::describe_posterior

For more, see- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# ----------------------- parametric ---------------------------------------

one_sample_test(
  data = ggplot2::msleep,
  x = brainwt,
  test.value = 0.275,
  type = "parametric"
)

#> # A tibble: 1 x 14
#>      mu statistic df.error p.value method            alternative estimate
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>             <chr>          <dbl>
#> 1 0.275    0.0504       55   0.960 One Sample t-test two.sided    0.00665
#>   conf.level conf.low conf.high effectsize conf.method conf.distribution
#>        <dbl>    <dbl>     <dbl> <chr>      <chr>       <chr>            
#> 1       0.95   -0.254     0.267 Hedges' g  ncp         t                
#>   expression
#>   <list>    
#> 1 <language>

# ----------------------- non-parametric -----------------------------------

one_sample_test(
  data = ggplot2::msleep,
  x = brainwt,
  test.value = 0.275,
  type = "nonparametric"
)

#> # A tibble: 1 x 11
#>   statistic    p.value method                    alternative estimate conf.level
#>       <dbl>      <dbl> <chr>                     <chr>          <dbl>      <dbl>
#> 1      5.51 0.00000738 Wilcoxon signed rank test two.sided     -0.689       0.95
#>   conf.low conf.high effectsize        conf.method expression
#>      <dbl>     <dbl> <chr>             <chr>       <list>    
#> 1   -0.817    -0.497 r (rank biserial) normal      <language>

# ----------------------- robust --------------------------------------------

one_sample_test(
  data = ggplot2::msleep,
  x = brainwt,
  test.value = 0.275,
  type = "robust"
)

#> # A tibble: 1 x 9
#>   statistic p.value method                                 estimate conf.low
#>       <dbl>   <dbl> <chr>                                     <dbl>    <dbl>
#> 1     -14.7       0 Bootstrap-t method for one-sample test   0.0390  0.00201
#>   conf.high conf.level effectsize   expression
#>       <dbl>      <dbl> <chr>        <list>    
#> 1    0.0759       0.95 Trimmed mean <language>

# ---------------------------- Bayesian -----------------------------------

one_sample_test(
  data = ggplot2::msleep,
  x = brainwt,
  test.value = 0.275,
  type = "bayes",
  bf.prior = 0.8
)

#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference  0.00851       0.95  -0.251      0.263 0.528          0.626 
#> 2 Cohens_d    0.276         0.95   0.0156     0.536 0.980          0.0576
#>   prior.distribution prior.location prior.scale  bf10 method          expression
#>   <chr>                       <dbl>       <dbl> <dbl> <chr>           <list>    
#> 1 cauchy                          0         0.8 0.130 Bayesian t-test <language>
#> 2 cauchy                          0         0.8 0.130 Bayesian t-test <language>

# }