statsExpressions: Tidy Dataframes and Expressions with Statistical Details

The {statsExpressions} package has two key aims:

to provide a consistent syntax to do statistical analysis with tidy data (in pipe-friendly manner),

to provide statistical expressions (pre-formatted in-text statistical results) for plotting functions.

Statistical packages exhibit substantial diversity in terms of their syntax and expected input type. This can make it difficult to switch from one statistical approach to another. For example, some functions expect vectors as inputs, while others expect dataframes. Depending on whether it is a repeated measures design or not, different functions might expect data to be in wide or long format. Some functions can internally omit missing values, while other functions error in their presence. Furthermore, if someone wishes to utilize the objects returned by these packages downstream in their workflow, this is not straightforward either because even functions from the same package can return a list, a matrix, an array, a dataframe, etc., depending on the function.

This is where {statsExpressions} comes in: It can be thought of as a unified portal through which most of the functionality in these underlying packages can be accessed, with a simpler interface and no requirement to change data format.

This package forms the statistical processing backend for ggstatsplot package.

For more documentation, see the dedicated website.

Details

statsExpressions

See also

Author

Maintainer: Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets) [copyright holder]