Expressions with statistics for tidy regression dataframes

Expressions with statistics for tidy regression dataframes

tidy_model_expressions(
  data,
  statistic = NULL,
  k = 2L,
  effsize.type = "omega",
  ...
)

Arguments

data
A tidy dataframe from regression model object.

statistic
Which statistic is to be displayed (either "t" or "f"or "z" or "chi") in the label.

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

effsize.type
Type of effect size needed for parametric tests. The argument can be "eta" (partial eta-squared) or "omega" (partial omega-squared).

...
Currently ignored.

Note

This is an experimental function and may change in the future. Please do not use it yet in your workflow.

Examples

set.seed(123)

# tidy dataframe
df <- tidy_model_parameters(lm(wt ~ am * cyl, mtcars))

# create a column containing expressions
tidy_model_expressions(df, statistic = "t")

#> # A tibble: 4 x 10
#>   term        estimate std.error conf.level conf.low conf.high statistic
#>   <chr>          <dbl>     <dbl>      <dbl>    <dbl>     <dbl>     <dbl>
#> 1 (Intercept)   1.66      0.587        0.95    0.455     2.86      2.82 
#> 2 am           -0.956     0.793        0.95   -2.58      0.668    -1.21 
#> 3 cyl           0.304     0.0826       0.95    0.135     0.473     3.68 
#> 4 am:cyl        0.0328    0.130        0.95   -0.234     0.300     0.252
#>   df.error  p.value
#>      <int>    <dbl>
#> 1       28 0.00864 
#> 2       28 0.238   
#> 3       28 0.000989
#> 4       28 0.803   
#>   label                                                                         
#>   <chr>                                                                         
#> 1 list(~widehat(italic(beta))=='1.66', ~italic(t)('28')=='2.82', ~italic(p)=='0…
#> 2 list(~widehat(italic(beta))=='-0.96', ~italic(t)('28')=='-1.21', ~italic(p)==…
#> 3 list(~widehat(italic(beta))=='0.30', ~italic(t)('28')=='3.68', ~italic(p)=='9…
#> 4 list(~widehat(italic(beta))=='0.03', ~italic(t)('28')=='0.25', ~italic(p)=='0…