Two-sample tests

A dataframe containing results from a two-sample test and effect size plus confidence intervals.

To see details about functions which are internally used to carry out these analyses, see the following vignette- https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Usage

two_sample_test(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  alternative = "two.sided",
  k = 2L,
  conf.level = 0.95,
  effsize.type = "g",
  var.equal = FALSE,
  bf.prior = 0.707,
  tr = 0.2,
  nboot = 100L,
  top.text = NULL,
  ...
)

Arguments

data

A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x

The grouping (or independent) variable from the dataframe data. In case of a repeated measures or within-subjects design, if subject.id argument is not available or not explicitly specified, the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted, the results can be inaccurate when there are more than two levels in x and there are NAs present. The data is expected to be sorted by user in subject-1,subject-2, ..., pattern.

y

The response (or outcome or dependent) variable from the dataframe data.

subject.id

Relevant in case of a repeated measures or within-subjects design (paired = TRUE, i.e.), it specifies the subject or repeated measures identifier. Important: Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

paired

Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

conf.level

Scalar between 0 and 1. If unspecified, the defaults return 95% confidence/credible intervals (0.95).

effsize.type

Type of effect size needed for parametric tests. The argument can be "d" (for Cohen's d) or "g" (for Hedge's g).

var.equal

a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

bf.prior

A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

nboot

Number of bootstrap samples for computing confidence interval for the effect size (Default: 100L).

top.text

Text to display on top of the Bayes Factor message. This is mostly relevant in the context of ggstatsplot package functions.

...

Currently ignored.

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# parametric -------------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x    = group,
  y    = extra,
  type = "p"
)
#> # A tibble: 1 × 18
#>   term  group mean.group1 mean.group2 statistic df.error p.value
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
#> 1 extra group        0.75        2.33     -1.86     17.8  0.0794
#>   method                  alternative effectsize estimate conf.level conf.low
#>   <chr>                   <chr>       <chr>         <dbl>      <dbl>    <dbl>
#> 1 Welch Two Sample t-test two.sided   Hedges' g    -0.796       0.95    -1.66
#>   conf.high conf.method conf.distribution n.obs expression  
#>       <dbl> <chr>       <chr>             <int> <list>      
#> 1    0.0919 ncp         t                    20 <expression>

# within-subjects design
two_sample_test(
  data       = dplyr::filter(bugs_long, condition %in% c("HDHF", "HDLF")),
  x          = condition,
  y          = desire,
  paired     = TRUE,
  subject.id = subject,
  type       = "p"
)
#> # A tibble: 1 × 16
#>   term   group     statistic df.error  p.value method        alternative
#>   <chr>  <chr>         <dbl>    <dbl>    <dbl> <chr>         <chr>      
#> 1 desire condition      3.61       89 0.000500 Paired t-test two.sided  
#>   effectsize estimate conf.level conf.low conf.high conf.method
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <chr>      
#> 1 Hedges' g     0.378       0.95    0.165     0.592 ncp        
#>   conf.distribution n.obs expression  
#>   <chr>             <int> <list>      
#> 1 t                    90 <expression>

# non-parametric ----------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x    = group,
  y    = extra,
  type = "np"
)
#> # A tibble: 1 × 14
#>   parameter1 parameter2 statistic p.value method                 alternative
#>   <chr>      <chr>          <dbl>   <dbl> <chr>                  <chr>      
#> 1 extra      group           25.5  0.0693 Wilcoxon rank sum test two.sided  
#>   effectsize        estimate conf.level conf.low conf.high conf.method n.obs
#>   <chr>                <dbl>      <dbl>    <dbl>     <dbl> <chr>       <int>
#> 1 r (rank biserial)    -0.49       0.95   -0.784   -0.0175 normal         20
#>   expression  
#>   <list>      
#> 1 <expression>

# within-subjects design
two_sample_test(
  data       = dplyr::filter(bugs_long, condition %in% c("HDHF", "HDLF")),
  x          = condition,
  y          = desire,
  paired     = TRUE,
  subject.id = subject,
  type       = "np"
)
#> # A tibble: 1 × 14
#>   parameter1 parameter2 statistic  p.value method                    alternative
#>   <chr>      <chr>          <dbl>    <dbl> <chr>                     <chr>      
#> 1 desire     condition       1796 0.000430 Wilcoxon signed rank test two.sided  
#>   effectsize        estimate conf.level conf.low conf.high conf.method n.obs
#>   <chr>                <dbl>      <dbl>    <dbl>     <dbl> <chr>       <int>
#> 1 r (rank biserial)    0.487       0.95    0.285     0.648 normal         90
#>   expression  
#>   <list>      
#> 1 <expression>

# robust ----------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x    = group,
  y    = extra,
  type = "r"
)
#> # A tibble: 1 × 11
#>   statistic df.error p.value
#>       <dbl>    <dbl>   <dbl>
#> 1      1.62     8.26   0.143
#>   method                                              
#>   <chr>                                               
#> 1 Yuen's test on trimmed means for independent samples
#>   effectsize                                              estimate conf.level
#>   <chr>                                                      <dbl>      <dbl>
#> 1 Algina-Keselman-Penfield robust standardized difference    -1.09       0.95
#>   conf.low conf.high n.obs expression  
#>      <dbl>     <dbl> <int> <list>      
#> 1    -4.68         0    20 <expression>

# within-subjects design
two_sample_test(
  data       = dplyr::filter(bugs_long, condition %in% c("HDHF", "HDLF")),
  x          = condition,
  y          = desire,
  paired     = TRUE,
  subject.id = subject,
  type       = "r"
)
#> # A tibble: 1 × 15
#>   statistic df.error p.value method                                            
#>       <dbl>    <dbl>   <dbl> <chr>                                             
#> 1      2.91       53 0.00528 Yuen's test on trimmed means for dependent samples
#>   effectsize                                              estimate conf.level
#>   <chr>                                                      <dbl>      <dbl>
#> 1 Algina-Keselman-Penfield robust standardized difference    0.410       0.95
#>   conf.low conf.high    mu small medium large n.obs expression  
#>      <dbl>     <dbl> <dbl> <dbl>  <dbl> <dbl> <int> <list>      
#> 1    0.255     0.592     0   0.1    0.3   0.5    90 <expression>

#' # Bayesian ------------------------------

# between-subjects design
two_sample_test(
  data = sleep,
  x    = group,
  y    = extra,
  type = "bayes"
)
#> # A tibble: 1 × 16
#>   term       effectsize      estimate conf.level conf.low conf.high    pd
#>   <chr>      <chr>              <dbl>      <dbl>    <dbl>     <dbl> <dbl>
#> 1 Difference Bayesian t-test    -1.13       0.95    -2.77     0.353 0.935
#>   rope.percentage prior.distribution prior.location prior.scale  bf10
#>             <dbl> <chr>                       <dbl>       <dbl> <dbl>
#> 1          0.0755 cauchy                          0       0.707  1.27
#>   method          log_e_bf10 n.obs expression  
#>   <chr>                <dbl> <int> <list>      
#> 1 Bayesian t-test      0.236    20 <expression>

# within-subjects design
two_sample_test(
  data       = dplyr::filter(bugs_long, condition %in% c("HDHF", "HDLF")),
  x          = condition,
  y          = desire,
  paired     = TRUE,
  subject.id = subject,
  type       = "bayes"
)
#> # A tibble: 1 × 16
#>   term       effectsize      estimate conf.level conf.low conf.high    pd
#>   <chr>      <chr>              <dbl>      <dbl>    <dbl>     <dbl> <dbl>
#> 1 Difference Bayesian t-test     1.09       0.95    0.471      1.69     1
#>   rope.percentage prior.distribution prior.location prior.scale  bf10
#>             <dbl> <chr>                       <dbl>       <dbl> <dbl>
#> 1               0 cauchy                          0       0.707  43.5
#>   method          log_e_bf10 n.obs expression  
#>   <chr>                <dbl> <int> <list>      
#> 1 Bayesian t-test       3.77    90 <expression>
# }